Allow templating repoUrl, revision/targetRevision in applicationsets (#9)
This fixes #2 and also aims to address some of the comments in https://github.com/orgs/k8tre/discussions/11

A developer wanting to deploy from a feature branch or wanting to deploy from a forked repo, needs to only change parameters in the `app_of_apps/root-app-of-apps.yaml` manifest and this will be patched through to all appsets.

This also replaces the directory source for the app_of_apps manifest to a Kustomize source. This means that any new appset manifests will only be discovered if they are added to `appsets/kustomization.yaml`.

This can be further improved with application level patching but there will be trade-offs between complexity and ease of use/maintenance.